% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{insectbraindb_species_info}
\alias{insectbraindb_species_info}
\alias{insectbraindb_neuron_info}
\title{Get information on species/neurons from insectbraindb.org}
\usage{
insectbraindb_species_info()

insectbraindb_neuron_info()
}
\value{
a \code{data.frame} that most importantly gives the user the latin scientific name, common name or neuron names and insectbraindb.org-specific ID number
for each insect species / neuron on which data is held on the site
}
\description{
Get information on the neurons and brain templates hosted by insectbraindb.org, including neuron names, species names, associated publications, etc.
}
\details{
As of May 2019, data from the following \href{https://insectbraindb.org/app/species}{species} is hosted on insectbraindb.org :
\itemize{
 \item 	\emph{Agrotis infusa}	 Bogong moth
 \item 	\emph{Agrotis segetum}	 Turnip moth
 \item 	\emph{Apis mellifera}	 Honeybee
 \item 	\emph{Apis mellifera}	 Honeybee
 \item 	\emph{Danaus plexippus}	 Monarch Butterfly
 \item 	\emph{Helicoverpa armigera}	 Cotton Bollworm,
 \item 	\emph{Helicoverpa assulta}	 Oriental tobacco budworm
 \item 	\emph{Heliothis virescens}	 Tobacco budworm
 \item 	\emph{Macroglossum stellatarum}	 Hummingbird hawk moth
 \item 	\emph{Manduca sexta}	 Tobacco hornworm
 \item 	\emph{Megalopta genalis}	 Sweat bee
 \item 	\emph{Nasonia vitripennis}	 Jewel wasp
 \item 	\emph{Scarabaeus lamarcki}	 Diurnal dung beetle
 \item 	\emph{Schistocerca gregaria}	 Desert Locust
 }
}
\seealso{
\code{\link{insectbraindb_read_brain}}, \code{\link{insectbraindb_read_neurons}}
}
