% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_brains.R
\name{insectbraindb_read_brain}
\alias{insectbraindb_read_brain}
\title{Read 3D insect brain meshes from insectbraindb.org}
\usage{
insectbraindb_read_brain(
  species = insectbraindb_species_info()$scientific_name,
  brain.sex = c("UNKNOWN", "MALE", "FEMALE"),
  progress = TRUE
)
}
\arguments{
\item{species}{the full scientific name for a species. The available options can be seen \href{https://insectbraindb.org/app/species}{here}}

\item{brain.sex}{the sex of the species' brain. The available options can be seen \href{https://insectbraindb.org/app/species}{here}}

\item{progress}{if \code{TRUE} or a numeric value, a progress bar is shown to track the state of your download}
}
\value{
a \code{nat} package \code{\link[nat]{hxsurf}} object, which mimics the Amira surface format, replete with metadata that can be
accessed using \code{$}
}
\description{
Read templatebrains, comprised of their different neuropils, for various insect species from from insectbraindb.org (2018 version),
given a single latin names for the species desired. Metadata for available neurons can be seen, and IDs chosen,
by calling \code{\link{insectbraindb_species_info}}. 3D triangular brain meshes are returned as a \code{nat} package \code{\link[nat]{hxsurf}}
object, which mimics the Amira surface format. These can be be plotted in 3D using \code{rgl} and analysed with tools from the \code{nat} ecosystem.
This includes subsetting by neuropil, i.e.. if you only want to visualise or analyse the antennal lobe.
}
\details{
A single 3D brain object is read, a .obj file for each of its neuropils is downloaded from https://ibdb-file-storage.s3.amazonaws.com/
 to a temporary directory, and read using \code{\link[readobj]{read.obj}} into a
 \code{\link[nat]{hxsurf}} object in R, which mimics the Amira surface format.
 As of May 2019, data from the following \href{https://insectbraindb.org/app/species}{species} is hosted on insectbraindb.org :
\itemize{
 \item 	\emph{Agrotis infusa}	 Bogong moth
 \item 	\emph{Agrotis segetum}	 Turnip moth
 \item 	\emph{Apis mellifera}	 Honeybee
 \item 	\emph{Apis mellifera}	 Honeybee
 \item 	\emph{Danaus plexippus}	 Monarch Butterfly
 \item 	\emph{Helicoverpa armigera}	 Cotton Bollworm,
 \item 	\emph{Helicoverpa assulta}	 Oriental tobacco budworm
 \item 	\emph{Heliothis virescens}	 Tobacco budworm
 \item 	\emph{Macroglossum stellatarum}	 Hummingbird hawk moth
 \item 	\emph{Manduca sexta}	 Tobacco hornworm
 \item 	\emph{Megalopta genalis}	 Sweat bee
 \item 	\emph{Nasonia vitripennis}	 Jewel wasp
 \item 	\emph{Scarabaeus lamarcki}	 Diurnal dung beetle
 \item 	\emph{Schistocerca gregaria}	 Desert Locust
 }
 Note that since neurons are reconstructed from many different neural species,
 there is no 'standard' orientation between species, but within a species these neurons are registered to a
 template brain, usually using elastix.
}
\seealso{
\code{\link{insectbraindb_neuron_info}}, \code{\link{insectbraindb_read_neurons}}, \code{\link{insectbraindb_species_info}}
}
