% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insectbrainr-package.R
\docType{package}
\name{insectbrainr-package}
\alias{insectbrainr}
\alias{insectbrainr-package}
\title{See \url{http://natverse.org/insectbrainr/} for structured package
  documentation.}
\description{
Provides programmatic access to the InsectBrainDB.org, which hosts
    over 100 neurons and template brains for over a dozen different insect 
    species. Principally curated by Stanley Heinze at the University of Lund, 
    contact him at: stanley.heinze@biol.lu.se.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/natverse/insectbrainr}
  \item Report bugs at \url{https://github.com/natverse/insectbrainr/issues}
}

}
\author{
\strong{Maintainer}: Alexander Bates \email{alexander.shakeel.bates@gmail.com} (\href{https://orcid.org/0000-0002-1195-0445}{ORCID})

}
\keyword{internal}
